requirements-base.txt
# Base requirements for all environments
# Python 3.12 compatible versions

# Core Framework
Django==5.1.4
djangorestframework==3.15.2

# Authentication
djangorestframework-simplejwt==5.4.0

# Database Drivers
psycopg2-binary==2.9.10

# CORS Headers
django-cors-headers==4.6.0

# File Upload & Media Processing
Pillow==11.0.0
python-magic==0.4.27

# Task Queue & Caching
celery==5.4.0
redis==5.2.1

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.1

# Security
django-ratelimit==4.1.0
cryptography==44.0.0

# API Documentation
drf-spectacular==0.28.0

# Utilities
django-filter==24.3
django-extensions==3.2.3

# Health Check
django-health-check==3.18.3
